+++ docs/CHANGELOG.rst
@@ -6,6 +6,15 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Changed
+-------
+- Support Python processes in Sphinx ``autoprocess*::`` directive
+
+
 ===================
 25.0.0 - 2020-10-16
 ===================